# app.js
/*********************************************************************************************************
 * Objetivo: Arquivo responsável pela criação do bando de dados sobre filmes
 * Data: 2026-04-17
 * Autor: Matheus Lucas
 * Versão: 1.0
 ********************************************************************************************************/

// Import das dependencias para criar a API
const express       = require('express')
const cors          = require('cors')
const bodyParser    = require('body-parser')

// Import das controllers do projeto
const controllerFilme           = require('./controller/filme/controller_filme.js')
const controllerGenero          = require('./controller/genero/controller_genero.js')
const controllerClassificacao   = require('./controller/classificacao/controllerClassificacao.js')
const controllerProfissional    = require('./controller/profissional/controllerProfissional.js')
const controllerNascionalidade  = require('./controller/nascionalidade/controllerNascionalidade.js')

// Permitindo a utilização do JSON no body das requisições.
const bodyParserJSON = bodyParser.json()


// Criando um objeto do express para criar a API
const app = express()

// Configurações do CORS da API
const corsOptions = {
    origin: ['*'], //Configuração de origem da requisição(IP ou Dominio).
    methods: 'GET, POST, PUT, DELETE, OPTIONS', //Configuração dos verbos que serão utilizados na API.
    allowedHeaders: ['Content-type', 'Authorization'] //Configurações de permissões.
                    //Tipo de dados  //Autorização de acesso 
}

// Aplica as configurações do CORS no app (EXPRESS)
app.use(cors(corsOptions))



// ENDPOINTS SOBRE O FILME

// endpoint para o CRUD do filme
app.post('/v1/senai/locadora/filme',bodyParserJSON, async function(request, response){
   // Recebendo o body da requisição
    let  dados = request.body


    // Recebendo o tipo de dados da requisição para validar se é JSON.
    let contentType = request.headers['content-type']

    let result = await controllerFilme.inserirNovoFilme(dados, contentType)

    response.status(result.status_code)
    response.json(result)
})

// endpoint para pegar todos os filmes
app.get('/v1/senai/locadora/filme', async function(request, response){
    let result = await controllerFilme.listarFilme()

    response.status(result.status_code)
    response.json(result)
})

// endpoint para pegar um filme pelo id
app.get('/v1/senai/locadora/filme/:id', async function(request, response) {
    let id = request.params.id

    let result = await controllerFilme.buscarFilme(id)

    response.status(result.status_code)
    response.json(result)
})

// endpoint para atualizar um filme
app.put('/v1/senai/locadora/filme/:id', bodyParserJSON, async function(request, response) {

    // Recebe o id do registro a ser atualizado.
    let id = request.params.id

    // Recebe os dados do body que serão modificados no banco de dados.
    let dados = request.body

    // Recebe o content-type da requisição para validar se é JSON.
    let contentType = request.headers['content-type']
    
    // Chama a função para atualizar o filme.
    let result = await controllerFilme.atualizarFilme(dados, id, contentType)

    response.status(result.status_code)
    response.json(result)
})

app.delete('/v1/senai/locadora/filme/:id', async function (request, response){
    let id = request.params.id

    let result = await controllerFilme.excluirFilme(id)

    response.status(result.status_code)
    response.json(result)
})



// ENDPOINTS SOBRE O GÊNERO

// endpoint para o CRUD do gênero
app.post('/v1/senai/locadora/genero',bodyParserJSON, async function(request, response){
    // Recebendo o body da requisição
     let  dados = request.body
 
 
     // Recebendo o tipo de dados da requisição para validar se é JSON.
     let contentType = request.headers['content-type']
 
     let result = await controllerGenero.inserirNovoGenero(dados, contentType)
 
     response.status(result.status_code)
     response.json(result)
 })

 // endpoint para listar os gêneros
 app.get('/v1/senai/locadora/genero', async function(request, response) {
    let result = await controllerGenero.listarGenero()

    response.status(result.status_code)
    response.json(result)
 })

 // endpoint para pegar um gênero pelo id
app.get('/v1/senai/locadora/genero/:id', async function(request, response) {
    let id = request.params.id

    let result = await controllerGenero.buscarGenero(id)

    response.status(result.status_code)
    response.json(result)
})

// endpoint para atualizar um filme
app.put('/v1/senai/locadora/genero/:id', bodyParserJSON, async function(request, response) {

    // Recebe o id do registro a ser atualizado.
    let id = request.params.id

    // Recebe os dados do body que serão modificados no banco de dados.
    let dados = request.body

    // Recebe o content-type da requisição para validar se é JSON.
    let contentType = request.headers['content-type']
    
    // Chama a função para atualizar o filme.
    let result = await controllerGenero.atualizarGenero(dados, id, contentType)

    response.status(result.status_code)
    response.json(result)
})

// endpoint para deletar gênero pelo id
app.delete('/v1/senai/locadora/genero/:id', async function(request, response) {
    let id = request.params.id

    let result = await controllerGenero.excluirGenero(id)
    response.status(result.status_code)
    response.json(result)
})



// ENDPOINTS SOBRE A CLASSIFICAÇÃO

// endpoint para CRUD da classificação
app.post('/v1/senai/locadora/classificacao',bodyParserJSON, async function(request, response){
    // Recebendo o body da requisição
     let  dados = request.body
 
 
     // Recebendo o tipo de dados da requisição para validar se é JSON.
     let contentType = request.headers['content-type']
 
     let result = await controllerClassificacao.inserirNovaClassificacao(dados, contentType)
 
     response.status(result.status_code)
     response.json(result)
 })

// endpoint para listar os classificação
app.get('/v1/senai/locadora/classificacao', async function(request, response) {
    let result = await controllerClassificacao.listarClassificacao()

    response.status(result.status_code)
    response.json(result)
})

// endpoint para listar por ID a classificação
app.get('/v1/senai/locadora/classificacao/:id', async function(request, response) {
    let id = request.params.id
    let result = await controllerClassificacao.buscarClassificacao(id)

    response.status(result.status_code)
    response.json(result)
})

// endpoint para atualizar uma classificação
app.put('/v1/senai/locadora/classificacao/:id', bodyParserJSON, async function(request, response) {

    // Recebe o id do registro a ser atualizado.
    let id = request.params.id

    // Recebe os dados do body que serão modificados no banco de dados.
    let dados = request.body

    // Recebe o content-type da requisição para validar se é JSON.
    let contentType = request.headers['content-type']
    
    // Chama a função para atualizar o filme.
    let result = await controllerClassificacao.atualizarClassificacao(dados, id, contentType)
    
    response.status(result.status_code)
    response.json(result)
})

// endpoint para deletar a classificação pelo id
app.delete('/v1/senai/locadora/classificacao/:id', async function(request, response) {
    let id = request.params.id

    let result = await controllerClassificacao.excluirClassificacao(id)
    response.status(result.status_code)
    response.json(result)
})



// ENDPOINTS SOBRE O PROFISSIONAL

// endpoint para inserir um profissional
app.post('/v1/senai/locadora/profissional',bodyParserJSON, async function(request, response){
    // Recebendo o body da requisição
     let  dados = request.body
 
 
     // Recebendo o tipo de dados da requisição para validar se é JSON.
     let contentType = request.headers['content-type']
 
     let result = await controllerProfissional.inserirNovoProfissional(dados, contentType)
 
     response.status(result.status_code)
     response.json(result)
 })

// endpoint para listar por ID o Profissonal
app.get('/v1/senai/locadora/profissional/:id', async function(request, response) {
    let id = request.params.id
    let result = await controllerProfissional.buscarProfissional(id)

    response.status(result.status_code)
    response.json(result)
})

// endpoint para atualizar um Profissional
app.put('/v1/senai/locadora/profissional/:id', bodyParserJSON, async function(request, response) {

    // Recebe o id do registro a ser atualizado.
    let id = request.params.id

    // Recebe os dados do body que serão modificados no banco de dados.
    let dados = request.body

    // Recebe o content-type da requisição para validar se é JSON.
    let contentType = request.headers['content-type']
    
    // Chama a função para atualizar o filme.
    let result = await controllerProfissional.atualizarProfissional(dados, id, contentType)
    
    response.status(result.status_code)
    response.json(result)
})

// endpoint para listar os Profissonais
app.get('/v1/senai/locadora/profissional', async function(request, response) {

    let result = await controllerProfissional.listarProfissional()
    response.status(result.status_code)
    response.json(result)
    
})

// endpoint para deletar algum profissional
app.delete('/v1/senai/locadora/profissional/:id', async function(request, response) {
    let id = request.params.id
    let result = await controllerProfissional.excluirProfissional(id)
    
    response.status(result.status_code)
    response.json(result)
})


// ENDPOINTS SOBRE A NASCIONALIDADE  

// endpoint para inserir a nascionalidade
app.post('/v1/senai/locadora/nascionalidade', bodyParserJSON, async function(request, response) {
   
    //Guardando os dados do body para utilizar no CRUD
    let dados = request.body

    let contentType = request.headers['content-type']

    let result = await controllerNascionalidade.inserirNovoNascionalidade(dados, contentType)

    response.status(result.status_code)
    response.json(result)

})

// endpoint para listar por ID a nascionalidade
app.get('/v1/senai/locadora/nascionalidade/:id', async function(request, response) {
    let id = request.params.id

    let result = await controllerNascionalidade.buscarNascionalidade(id)

    response.status(result.status_code)
    response.json(result)

})

// endpoint para atualizar uma nascionalidade
app.put('/v1/senai/locadora/nascionalidade/:id', bodyParserJSON, async function(request, response) {

    // Recebe o id do registro a ser atualizado.
    let id = request.params.id

    // Recebe os dados do body que serão modificados no banco de dados.
    let dados = request.body

    // Recebe o content-type da requisição para validar se é JSON.
    let contentType = request.headers['content-type']
    
    // Chama a função para atualizar o filme.
    let result = await controllerNascionalidade.atualizarNascionalidade(dados, id, contentType)
    
    response.status(result.status_code)
    response.json(result)
})

// endpoint para listar as nascionalidades
app.get('/v1/senai/locadora/nascionalidade', async function(request, response) {

    let result = await controllerNascionalidade.listarNascionalidade()
    response.status(result.status_code)
    response.json(result)
    
})

// endpoint para deletar algum profissional
app.delete('/v1/senai/locadora/nascionalidade/:id', async function(request, response) {
    let id = request.params.id
    let result = await controllerNascionalidade.excluirNascionalidade(id)
    
    response.status(result.status_code)
    response.json(result)
})



// Faz um start na API (Aguardando requisição)
app.listen(8080, function(){
    console.log('API aguardando novas requisições...')
})