# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto controle de Filmes - SENAI
  description: |-
   API destinada ao controle de Filmes, para o projeto "Bem Legal" do SENAI
  termsOfService: http://swagger.io/terms/
  contact:
    email: matheuslucas987123@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.4.26
externalDocs: 
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://
paths:
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo Filme
      description: Endpoint para enviar os dados do filme a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo Filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '415':
          description: Invalid Content-Type
        '500':
          description: Error Internal server (CONTROLLER/MODEL)
    get:
      tags:
        - Filme
      summary: Lista todos os Filmes
      description: Listagem de todos os filmes que foram cadastrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'          
        '404':
          description: Not Found
        '500':
          description: Error Internal server (CONTROLLER/MODEL)
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Busca um Filme pelo ID
      description: Retorna um Filme utilizando o ID do mesmo
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do Filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (CONTROLLER/MODEL)

    put:
      tags:
        - Filme
      summary: Atualizando um Filme existente
      description: Buscar e atualizar um filme utilizando o ID
      parameters:
        - name: id
          in: path
          description: ID of pet to return
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um Filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '415':
          description: Invalid content-Type
        '500':
          description: Error Internal Server(CONTROLLER/MODEL)    

    delete:
      tags:
        - Filme
      summary: Delete Filme
      description: Endpoint para deletar um Filme existente utilizando o ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (CONTROLLER/MODEL)
components:
  schemas:
    Filme:
      required:
        - nome
        - sinopse
        - data_lancamento
        - duracao
        - capa
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sinopse:
          type: string
        data_lancamento:
          type: date
          exemple: "2026-01-01"
        duracao:
          type: time
          exemple: "01:00:00"
        capa:
          type: string
        valor:
          type: decimal
          exemple: "10.5"
        avaliacao: 
          type: integer
  requestBodies:
    Filme:
      description: Objeto que deve ser enviado no Body das requisicoes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Filme'
    